relocate: Mypage 폴더 안에 있던 hooks를 바깥으로 꺼냄
