+++ .eslintrc
@@ -7,5 +7,8 @@
     "react/prop-types": "off",
     "react/button-has-type": "off",
     "comma-dangle": "off",
-  },
+    "prettier/prettier": ["error", { "singleQuote": false }],
+    "jsx-a11y/click-events-have-key-events": "off",
+    "jsx-a11y/no-static-element-interactions": "off"
+  }
 }
